requirements.txt
pyyaml>=6.0
flask>=2.0.0
bleak>=0.20.0
adafruit-circuitpython-neopixel>=6.3.0
RPi.GPIO>=0.7.0
cryptography>=3.4.8
